requirements.txt
# Core
python-dotenv>=1.0.1
pydantic>=2.7
pydantic-settings>=2.4
typer>=0.12
rich>=13.7
numpy>=1.26
tenacity>=8.2.0
PyYAML>=6.0

# HTTP API
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
PyJWT>=2.9
SQLAlchemy>=2.0
psycopg[binary]>=3.2
boto3>=1.35
azure-storage-blob>=12.23

# PDF -> images
pymupdf>=1.24.9
Pillow>=10.3

# OpenAI
openai>=1.45.0

# Testing
pytest>=8.3
